pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sql-manager"
version = "1.0.0"
description = "SQL数据库管理工具"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}
authors = [
    {name = "Al-say", email = "your-email@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

# 基础依赖（必需）
dependencies = []

[project.optional-dependencies]
# 完整功能依赖
full = [
    "SQLAlchemy>=1.4.0",
    "PyMySQL>=1.0.0",
    "psycopg2-binary>=2.9.0",
    "pyodbc>=4.0.30",
    "PyYAML>=6.0",
    "pandas>=1.3.0",
    "numpy>=1.21.0",
    "matplotlib>=3.4.0",
    "plotly>=5.0.0",
    "seaborn>=0.11.0",
    "python-dotenv>=0.19.0",
    "click>=8.0.0",
    "rich>=10.0.0",
    "tabulate>=0.8.9",
    "cryptography>=3.4.0",
    "python-dateutil>=2.8.0",
]

# 数据库支持
database = [
    "SQLAlchemy>=1.4.0",
    "PyMySQL>=1.0.0",
    "psycopg2-binary>=2.9.0",
    "pyodbc>=4.0.30",
]

# 数据分析支持
analysis = [
    "pandas>=1.3.0",
    "numpy>=1.21.0",
    "matplotlib>=3.4.0",
    "plotly>=5.0.0",
    "seaborn>=0.11.0",
]

# 开发依赖
dev = [
    "pytest>=6.2.0",
    "pytest-cov>=2.12.0",
    "black>=21.0.0",
    "flake8>=3.9.0",
    "mypy>=0.910",
]

[project.urls]
Homepage = "https://github.com/Al-say/SQL"
Repository = "https://github.com/Al-say/SQL.git"
Issues = "https://github.com/Al-say/SQL/issues"

[project.scripts]
sql-manager = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-data]
"*" = ["*.md", "*.txt", "*.yml", "*.yaml", "*.json"]

[tool.black]
line-length = 88
target-version = ['py37']

[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "--cov=src --cov-report=html"
